Move Frequency enum to own file
