+++ habit_task_tracker/lib/frequency.dart
@@ -1,0 +1,9 @@
+enum Frequency { daily, weekly, monthly, yearly, none }
+
+Map<String, Frequency> frequencyMap = {
+  'Frequency.daily': Frequency.daily,
+  'Frequency.weekly': Frequency.weekly,
+  'Frequency.monthly': Frequency.monthly,
+  'Frequency.yearly': Frequency.yearly,
+  'Frequency.none': Frequency.none,
+};
