+++ habit_task_tracker/lib/notifier.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/foundation.dart';
 import 'package:habit_task_tracker/habit.dart';
+import 'package:habit_task_tracker/frequency.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_local_notifications/flutter_local_notifications.dart';
 import 'package:timezone/data/latest_all.dart' as tz;
