+++ habit_task_tracker/test/log_test.dart
@@ -1,6 +1,6 @@
 import 'package:flutter_test/flutter_test.dart';
 import 'package:habit_task_tracker/log.dart';
-import 'package:habit_task_tracker/habit.dart';
+import 'package:habit_task_tracker/frequency.dart';
 import 'package:habit_task_tracker/uuid.dart';
 import '_setup_mocks.dart';
